backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
openai==1.54.4
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
langchain_community==0.3.8
langchain_openai==0.2.9
psycopg2-binary==2.9.10